Eliminate libdes dependency
We were only using libdes for des_read_pw_string().  To eliminate that
dependency, import a copy of libhcrypto's UI_UTIL_read_pw_string(), which has
the same interface.  I would just link to libhcrypto, but that library only
exists in newer heimdal.  In the heimdal ~0.6 we currently use in MiniWorld,
that library is called libdes, which is bound to cause confusion.