unlock_kdb: try multiple data files
Instead of attempting to use only one data file for a given user,
we now examine all files in the data directory with names belonging
to a family derived from the username.  This allows the use of
multiple tokens per user, which turns out to be important during
transition to a new token system.